General frontend refactor (#211)
* removes anything advising related #209

the courses soffit doesn't have advising baked in anymore, so anything
advising related in the frontend is no longer needed

* removes excess reducers/actions #209

the soffit pulls data in differently, so some actions/reducers
ended up not being needed in the end

* removes any use of useTranslation #209

translations breaks in portlet to soffit conversions, so its getting
ripped out for now, until we can find a better way to do it. other
commits might also include translations stuff

* changes buildingRoom to location #209

* updates reducers, actions, and api.js #209

* adds pdf service #209

the pdf service was ripped out of the backend. this is its replacement

* general refactor #209

commits the rest of the general refactor, where most of the state
changes are happening

* updates deps #209

* modifications to print element #209

* improves styling and flow of state #209

* updates error card & adds it to grade section #209

* fix styling bugs/remove unused props #209

* replaces local backend with demo data #209

* changes to 3rd party calendar, modify actions #209

* misc styling/cleaning up #209

* updates pdf service #209

pdf generation has been moved back to the backend

* fix anchorEl bug in popover #209

* add/run prettier, remove unused code #209

* updates readme, turns demo data off #209

* updates demo data per @nickwoodle suggestion #209

also patches a bug found after changing a class's demo data to have
no meetings

* replaces oakland domain with test domain #209

* change pdf endpoint wording #209

the endpoint changed in the backend so i have to change it here too

* removes unneeded deps #209